requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.0.0
notebook>=7.0.0
polars>=0.18.0
pyarrow>=12.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
tqdm>=4.65.0
joblib>=1.3.0
dask>=2023.6.0
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
invoke>=2.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
loguru>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0
transformers>=4.30.0

# Statistics and advanced analytics
# statsmodels>=0.14.0
# pingouin>=0.5.0
# lifelines>=0.27.0

# Computer vision
# opencv-python>=4.8.0
# Pillow>=10.0.0
# albumentations>=1.3.0

# Natural language processing
# nltk>=3.8.0
# spacy>=3.6.0
# transformers>=4.30.0
# gensim>=4.3.0

# Model deployment and serving
# mlflow>=2.5.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
# streamlit>=1.25.0

# Data validation and quality
# pydantic>=2.0.0
# great-expectations>=0.17.0
# pandera>=0.15.0

# Hyperparameter optimization
# optuna>=3.2.0
# hyperopt>=0.2.7
# scikit-optimize>=0.9.0

# Time series analysis
# prophet>=1.1.0
# pmdarima>=2.0.0
# tslearn>=0.6.0